send word to mdict

+++ include/ocr.h
@@ -9,6 +9,6 @@ namespace winrt::Windows::Graphics::Imaging {
 
 
 // todo: make as class
-/// @param distance_threshold maximum tolerate manhattan distance
+/// @param distance_threshold maximum tolerate euclidean distance
 std::wstring ocr(const winrt::Windows::Graphics::Imaging::SoftwareBitmap& sbm,
-                 const int distance_threshold);
+                 const float distance_threshold);
